requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
html5lib==1.0.1
idna==2.8
Pillow==5.4.1
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
webencodings==0.5.1
